# Optimise PEX by avoiding intermediate storage