Turns out, storing the PEX message entries in maps for later extraction at message generation stage is very expensive. If we prepare the lists without storing in maps we gain a huge win for the initial message. Here are the new benchmarks:

```
BenchmarkPexInitial4-8     	  217302	      5089 ns/op
BenchmarkPexInitial50-8    	   24865	     48012 ns/op
BenchmarkPexInitial100-8   	   11413	     90340 ns/op
BenchmarkPexInitial200-8   	    6747	    189316 ns/op
BenchmarkPexInitial400-8   	    2803	    438807 ns/op
```

Comparing these against the "after" benchmark of #435, we may observe significant improvement:

<img width="369" alt="Screenshot 2020-11-12 at 21 38 26" src="https://user-images.githubusercontent.com/11537279/98999270-67cb4f00-252f-11eb-9823-a032144332fc.png">

Accounting for reconnections may become slower but those, at later stages, come on top of the log processed earlier, amortising the cost.